pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "code-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["anthropic", "pyyaml"]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy", "types-PyYAML"]

[project.scripts]
agent = "agent.__main__:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
strict = true
